refactor: auth 도메인 추가 후 회원가입 로직 이동
