+++ src/main/java/until/the/eternity/das/auth/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package until.the.eternity.das.user.dto.request;
+package until.the.eternity.das.auth.dto.request;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
