+++ src/main/java/until/the/eternity/das/auth/dto/response/SignUpResponse.java
@@ -1,4 +1,4 @@
-package until.the.eternity.das.user.dto.response;
+package until.the.eternity.das.auth.dto.response;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
