+++ src/main/java/until/the/eternity/das/auth/exception/NicknameAlreadyExistsException.java
@@ -1,6 +1,6 @@
-package until.the.eternity.das.user.exception;
+package until.the.eternity.das.auth.exception;
 
-import static until.the.eternity.das.user.exception.UserExceptionCode.NICKNAME_ALREADY_EXISTS;
+import static until.the.eternity.das.auth.exception.AuthExceptionCode.NICKNAME_ALREADY_EXISTS;
 
 import until.the.eternity.das.common.exception.CustomException;
